refactor: change home page section layout
