--- src/views/Home/About/index.vue
@@ -1,92 +1,0 @@
-<template>
-  <div class="about">
-    <div class="about__banner" v-lazy-container="{ selector: 'img' }">
-      <img
-        data-src="https://storage.googleapis.com/openhci2021-storage/home-page/home_banner_720w.png"
-        data-srcset="
-          https://storage.googleapis.com/openhci2021-storage/home-page/home_banner_1440.png 1440w,
-          https://storage.googleapis.com/openhci2021-storage/home-page/home_banner_1440.png  960w,
-          https://storage.googleapis.com/openhci2021-storage/home-page/home_banner_1440.png  720w
-        "
-        sizes="(max-width: 768px) 400px, (max-width: 968px) 500px, 100vw"
-        data-loading="https://storage.googleapis.com/openhci2021-storage/home-page/home_banner_loading.png"
-        alt="OpneHCI 2021 Home Page"
-        class="about__banner--large-image"
-      />
-      <img
-        data-src="https://storage.googleapis.com/openhci2021-storage/home-page/home_banner_320.png"
-        alt="OpenHCI 2021 Home Page"
-        class="about__banner--small-image"
-      />
-    </div>
-    <Mobius />
-    <OpenHci />
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import Mobius from "./Mobius";
-import OpenHci from "./OpenHci";
-
-export default {
-  name: "About",
-  components: {
-    Mobius,
-    OpenHci,
-  },
-};
-</script>
-
-<style lang="scss" rel="stylesheet/scss" scoped>
-$md: 768px;
-$sm: 576px;
-
-.about {
-  &__banner {
-    background-color: #0a09fa;
-    position: relative;
-    padding-top: 57%;
-    margin-bottom: 174px;
-
-    @media (max-width: 768px) {
-      margin-bottom: 128px;
-    }
-
-    @media (max-width: 576px) {
-      padding-top: 177%;
-      margin-bottom: 64px;
-    }
-
-    img {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-    }
-
-    &--large-image {
-      @media (max-width: 576px) {
-        display: none;
-      }
-    }
-
-    &--small-image {
-      display: none;
-
-      @media (max-width: 576px) {
-        display: block;
-      }
-    }
-  }
-}
-
-.v-lazy-image {
-  filter: blur(10px);
-  transition: filter 0.7s;
-}
-
-.v-lazy-image-loaded {
-  filter: blur(0);
-}
-</style>
