+++ src/views/Home/Taichi/index.vue
